changed joystick motion value to be a float

+++ include/event.h
@@ -40,7 +40,7 @@ typedef struct ButtonUpEvent {
 typedef struct JoystickMotionEvent {
     JoystickType joystick;
     JoystickAxis axis;
-    int value;
+    float value;
 } JoystickMotionEvent;
 
 typedef void* Event;
